change_type_2

+++ src/main.java
@@ -2,7 +2,7 @@ public class main {
     public static void main(String[] args) {
         int balance = 2_000_000_000;
         int payment = 500_000_000;
-        long total = balance + payment;
+        int total = balance + payment;
         System.out.println(total);
 
     }
